Add the Map screen prototype

+++ app/src/main/java/com/sixt/task/model/vo/Car.kt
@@ -42,10 +42,10 @@ data class Car(
     val licensePlate: String,
 
     @SerializedName("latitude")
-    val latitude: BigDecimal,
+    val latitude: Double,
 
     @SerializedName("longitude")
-    val longitude: BigDecimal,
+    val longitude: Double,
 
     @SerializedName("innerCleanliness")
     val innerCleanliness: String,
